+++ irtt.go
@@ -43,6 +43,8 @@ func getCommand(name string) *command {
 func init() {
 	registerCommand("client", "runs the client", runClientCLI, clientUsage)
 	registerCommand("server", "runs the server", runServerCLI, serverUsage)
+	registerCommand("report", "emits results from a JSON file", runReport,
+		reportUsage)
 	registerCommand("bench", "runs HMAC and fill benchmarks", runBench, nil)
 	registerCommand("timer", "runs timer resolution test", runTimer, nil)
 	registerCommand("clock", "runs wall vs monotonic clock test", runClock, nil)