+++ irtt_client.go
@@ -341,7 +341,7 @@ func runClientCLI(args []string) {
 
 	// print results
 	if !*reallyQuiet {
-		printResult(r)
+		printResult(&r.PrintableResult)
 	}
 
 	// write results to JSON
@@ -352,7 +352,7 @@ func runClientCLI(args []string) {
 	}
 }
 
-func printResult(r *Result) {
+func printResult(r *PrintableResult) {
 	// set some stat variables for later brevity
 	rtts := r.RTTStats
 	rttvs := r.RoundTripIPDVStats
@@ -364,15 +364,20 @@ func printResult(r *Result) {
 	tes := r.TimerErrorStats
 	sps := r.ServerProcessingTimeStats
 
+	var e bool
 	if r.SendErr != nil {
 		if r.SendErr != context.Canceled {
 			printf("\nTerminated due to send error: %s", r.SendErr)
+			e = true
 		}
 	}
 	if r.ReceiveErr != nil {
 		printf("\nTerminated due to receive error: %s", r.ReceiveErr)
+		e = true
 	}
-	printf("")
+	if e {
+		printf("")
+	}
 
 	printStats := func(title string, s DurationStats) {
 		if s.N > 0 {
@@ -419,7 +424,6 @@ func printResult(r *Result) {
 	}
 	printf("     bytes sent/received: %d/%d", r.BytesSent, r.BytesReceived)
 	printf("       send/receive rate: %s / %s", r.SendRate, r.ReceiveRate)
-	printf("           packet length: %d bytes", r.Config.Length)
 	printf("             timer stats: %d/%d (%.2f%%) missed, %.2f%% error",
 		r.TimerMisses, r.ExpectedPacketsSent, r.TimerMissPercent,
 		r.TimerErrPercent)